backend/requirements.txt
flask==2.3.3
tensorflow==2.13.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
flask-cors==4.0.0
werkzeug==2.3.7